feat: expose connector catalog icon descriptors
